chore : 파일 구조 변경
error page 생성, views 하위폴더 생성
