+++ src/components/Layout/layout.css
@@ -1,0 +1,6 @@
+.layout {
+}
+.headerContainer {
+  display: flex;
+  flex-direction: row;
+}
