--- src/views/about.ts
@@ -1,6 +1,0 @@
-import Layout from "../components/Layout/layout";
-
-export default function About() {
-  const content = `<div>소개 페이지 컨텐츠</div>`;
-  return Layout(content);
-}
