+++ src/views/about/about.css
@@ -1,0 +1,2 @@
+.aboutContainer {
+}
