+++ src/views/error/error.css
@@ -1,0 +1,5 @@
+.asd {
+  background-color: red;
+}
+.errorText {
+}
