--- src/views/guestBook.ts
@@ -1,6 +1,0 @@
-import Layout from "../components/Layout/layout";
-
-export default function GuestBook() {
-  const content = `<div>방명록 페이지</div>`;
-  return Layout(content);
-}
