+++ src/views/guestBook/guestBook.ts
@@ -1,0 +1,6 @@
+import Layout from "../../components/Layout/layout";
+import styles from "./guestBook.css";
+export default function GuestBook() {
+  const content = `<div class=${styles.guestBookContainer}>방명록 페이지</div>`;
+  return Layout(content);
+}
