--- src/views/home.ts
@@ -1,6 +1,0 @@
-import Layout from "../components/Layout/layout";
-
-export default function Home() {
-  const content = `<div>홈 페이지 컨텐츠</div>`;
-  return Layout(content);
-}
