+++ src/views/home/home.css
@@ -1,0 +1,2 @@
+.homeContainer {
+}
