+++ src/views/post/post.css
@@ -1,0 +1,2 @@
+.postContainer {
+}
